Restructure App Startup
