+++ code/backend/app.py
@@ -1,72 +1,43 @@
-import sys
-import traceback
-from datetime import datetime
-
 from aiohttp import web
 from aiohttp.web import Request, Response
-from botbuilder.core import TurnContext
 from botbuilder.core.integration import aiohttp_error_middleware
 from botbuilder.integration.aiohttp import (
     CloudAdapter,
     ConfigurationBotFrameworkAuthentication,
 )
-from botbuilder.schema import Activity, ActivityTypes
-from bots.assistant import AssistantBot
+from bots.assistant_bot import AssistantBot
+from bots.utils_bot import BotUtils
 from core.config import settings as CONFIG
 from utils import enable_logging
 
-# Enable logging
-enable_logging()
 
-# Create adapter.
-# See https://aka.ms/about-bot-adapter to learn more about how bots work.
-ADAPTER = CloudAdapter(ConfigurationBotFrameworkAuthentication(CONFIG))
+def init_app() -> web.Application:
+    # Create cloud adapter
+    adapter = CloudAdapter(ConfigurationBotFrameworkAuthentication(CONFIG))
+    adapter.on_turn_error = BotUtils.on_error
 
+    # Create bot
+    bot = AssistantBot()
 
-# Catch-all for errors.
-async def on_error(context: TurnContext, error: Exception):
-    # This check writes out errors to console log .vs. app insights.
-    # NOTE: In production environment, you should consider logging this to Azure
-    #       application insights.
-    print(f"\n [on_turn_error] unhandled error: {error}", file=sys.stderr)
-    traceback.print_exc()
+    # Create app
+    app = web.Application(middlewares=[aiohttp_error_middleware])
 
-    # Send a message to the user
-    await context.send_activity("The bot encountered an error or bug.")
-    await context.send_activity(
-        "To continue to run this bot, please fix the bot source code."
-    )
-    # Send a trace activity if we're talking to the Bot Framework Emulator
-    if context.activity.channel_id == "emulator":
-        # Create a trace activity that contains the error object
-        trace_activity = Activity(
-            label="TurnError",
-            name="on_turn_error Trace",
-            timestamp=datetime.now(datetime.UTC),
-            type=ActivityTypes.trace,
-            value=f"{error}",
-            value_type="https://www.botframework.com/schemas/error",
-        )
-        # Send a trace activity, which will be displayed in Bot Framework Emulator
-        await context.send_activity(trace_activity)
+    # Listen for incoming requests on /api/messages
+    async def messages(req: Request) -> Response:
+        return await adapter.process(req, bot)
 
+    # Add route to app
+    app.router.add_post("/api/messages", messages)
 
-ADAPTER.on_turn_error = on_error
-
-# Create the Bot
-BOT = AssistantBot()
-
-
-# Listen for incoming requests on /api/messages
-async def messages(req: Request) -> Response:
-    return await ADAPTER.process(req, BOT)
-
-
-APP = web.Application(middlewares=[aiohttp_error_middleware])
-APP.router.add_post("/api/messages", messages)
 
 if __name__ == "__main__":
+    # Enable logging
+    enable_logging()
+
+    # Create app
+    app = init_app()
+
     try:
-        web.run_app(APP, host="localhost", port=CONFIG.PORT)
+        web.run_app(app, host="localhost", port=CONFIG.PORT)
     except Exception as error:
         raise error
