+++ code/backend/bots/assistant_bot.py
